Fixed dependency cycles
