+++ package.json
@@ -35,18 +35,18 @@
     "wasgen": "^0.18.2"
   },
   "devDependencies": {
-    "@typescript-eslint/parser": "^5.53.0",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
+    "@typescript-eslint/parser": "^5.53.0",
     "cross-env": "^7.0.3",
     "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-app": "^6.2.2",
+    "eslint-config-react-app": "^7.0.1",
     "husky": "^8.0.2",
     "lint-staged": "^13.0.3",
     "prettier": "^2.8.4",
     "prettier-plugin-organize-imports": "^3.0.0",
-    "react-scripts-plugin-no-minify": "^0.4.0",
     "wasm-loader": "^1.3.0"
   },
   "husky": {
