+++ src/features/gameboard/EmergencyStop/EmergencyStop.tsx
@@ -4,8 +4,8 @@ import { Icon } from '../../../helpers/Icon'
 
 import { type AnyAction } from 'redux'
 import { type ThunkDispatch } from 'redux-thunk'
+import { loop, wait } from '../../../helpers/helpers'
 import { type IState } from '../../../store'
-import { loop, wait } from '../events/helpers'
 import { MessageType } from '../MessageArea/MessageTypes'
 import { playTone } from '../sound'
 import { GameBoardActions } from '../store'
