+++ src/features/gameboard/index.ts
@@ -1,2 +1,3 @@
+export * from './events'
 export { GameBoard } from './GameBoard'
 export * from './store'
