+++ src/features/gameboard/store/events/index.ts
@@ -1,0 +1,9 @@
+export { Cum } from './edging/cum'
+export { Edge } from './edging/edge'
+export { DoublePace } from './pace/doublePace'
+export { HalfPace } from './pace/halfPace'
+export { Pause } from './pace/pause'
+export { RandomPace } from './pace/randomPace'
+export { RisingPace } from './pace/risingPace'
+export { CleanMess } from './tasks/cleanMess'
+export { RandomGrip } from './tasks/randomGrip'
