+++ src/features/gameboard/store/index.ts
@@ -1,1 +1,2 @@
+export * from './actions'
 export * from './reducer'
