+++ src/pages/Greeter/Greeter.css
@@ -47,6 +47,7 @@
   height: auto;
   font-weight: bold;
   padding: 15px 20px;
+  margin: 20px 0px;
 }
 
 .GreeterPage a.no-underline {
